refactor: create separate component for typewriter effect
